+++ src/application.rs
@@ -70,6 +70,7 @@ use crate::checkers::wp_plugins::elements_ready_lite::ElementsReadyLiteChecker;
 use crate::checkers::wp_plugins::email_subscribers::EmailSubscribersChecker;
 use crate::checkers::wp_plugins::forminator::ForminatorChecker;
 use crate::checkers::wp_plugins::gtranslate::GTranslateChecker;
+use crate::checkers::wp_plugins::health_check::HealthCheckChecker;
 use crate::checkers::wp_plugins::jetpack::JetpackChecker;
 use crate::checkers::wp_plugins::js_composer::JSComposerChecker;
 use crate::checkers::wp_plugins::layerslider::LayerSliderChecker;
@@ -162,6 +163,7 @@ impl Application {
             Box::new(ForminatorChecker::new()),
             Box::new(DiviChecker::new()),
             Box::new(GTranslateChecker::new()),
+            Box::new(HealthCheckChecker::new()),
             Box::new(JetpackChecker::new()),
             Box::new(JSComposerChecker::new()),
             Box::new(LayerSliderChecker::new()),