+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Detection for Email Subscribers WordPress plugin.
 - Detection for Better Search Replace WordPress plugin.
 - Detection for Advanced Custom Fields WordPress plugin
+- Detection for Health Check & Troubleshooting WordPress plugin.
 
 ### Fixed